added authServiceImpl, removed unused imports
